gitlab-mirrors-api/requirements.txt
click==8.1.3
colorama==0.4.6
configparser==5.3.0
Flask==2.2.2
gitdb==4.0.10
GitPython==3.1.30
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
smmap==5.0.0
typing-extensions==4.4.0
Werkzeug==2.2.2
zipp==3.11.0
